+++ mne/__init__.py
@@ -49,8 +49,8 @@ from .forward import (read_forward_solution, apply_forward, apply_forward_raw,
 from .source_estimate import (read_source_estimate, MixedSourceEstimate,
                               SourceEstimate, VectorSourceEstimate,
                               VolSourceEstimate, morph_data,
-                              morph_data_precomputed, compute_morph_matrix,
-                              grade_to_tris, grade_to_vertices,
+                              morph_data_precomputed,
+                              grade_to_tris,
                               spatial_src_connectivity,
                               spatial_tris_connectivity,
                               spatial_dist_connectivity,
@@ -59,7 +59,8 @@ from .source_estimate import (read_source_estimate, MixedSourceEstimate,
                               spatio_temporal_tris_connectivity,
                               spatio_temporal_dist_connectivity,
                               save_stc_as_volume, extract_label_time_course)
-from .morph import SourceMorph, read_source_morph
+from .morph import (SourceMorph, read_source_morph, compute_morph_matrix,
+                    grade_to_vertices)
 from .surface import (read_surface, write_surface, decimate_surface, read_tri,
                       read_morph_map, get_head_surf, get_meg_helmet_surf)
 from .source_space import (read_source_spaces, vertex_to_mni,
