+++ mne/morph.py
@@ -5,16 +5,23 @@
 
 import os.path as op
 import warnings
+import copy
+from math import ceil
 
 import numpy as np
+from scipy import sparse
+from scipy.sparse import block_diag as sparse_block_diag
 
-from . import read_forward_solution, compute_morph_matrix
+from .forward import read_forward_solution
+from .parallel import parallel_func
+from .source_estimate import (VolSourceEstimate, SourceEstimate,
+                              VectorSourceEstimate, _get_ico_tris)
+from .source_space import SourceSpaces
+from .surface import read_morph_map, mesh_edges, read_surface, _compute_nearest
+from .utils import (logger, verbose, check_version, get_subjects_dir,
+                    _check_subject, warn as warn_, deprecated)
 from .externals.h5io import read_hdf5, write_hdf5
 from .externals.six import string_types
-from .source_estimate import (VolSourceEstimate, SourceEstimate,
-                              VectorSourceEstimate)
-from .source_space import SourceSpaces
-from .utils import logger, verbose, check_version
 
 
 class SourceMorph(object):
@@ -51,7 +58,15 @@ class SourceMorph(object):
     xhemi : bool
         Morph across hemisphere. Currently only implemented for
         ``subject_to == subject_from``. See notes below.
-
+    spacing : int | list | None
+        int, list (of two arrays), or None
+        Resolution of the icosahedral mesh (typically 5). If None, all
+        vertices will be used (potentially filling the surface). If a list,
+        then values will be morphed to the set of vertices specified in
+        in spacing[0] and spacing[1]. Note that specifying the vertices (e.g.,
+        spacing=[np.arange(10242), np.arange(10242)] for fsaverage on a
+        standard spacing 5 source space) can be substantially faster than
+        computing vertex locations. Default is spacing=5
 
     Volume Source Estimate parameters
     ---------------------------------
@@ -66,11 +81,11 @@ class SourceMorph(object):
         Increasing index values for the tuple mean later levels and
         each int represents the number of iterations in that level. Default is
         niter_sdr=(5, 5, 3)
-    grid_spacing : tuple
+    spacing : tuple | int | float | None
         Voxel size of volume for each spatial dimension in mm.
-        If grid_spacing is None, MRIs won't be resliced. Note that in this case
+        If spacing is None, MRIs won't be resliced. Note that in this case
         both volumes must have the same number of slices in every
-        spatial dimension. Default is grid_spacing=(5., 5., 5.)
+        spatial dimension. Default is spacing=(5., 5., 5.)
 
     Attributes
     ----------
@@ -97,11 +112,21 @@ class SourceMorph(object):
         iterations to perform the sdr transform. Increasing index values
         for the tuple mean later levels and each int represents the number
         of iterations in that level.
-    grid_spacing : tuple
-        Voxel size of volume for each spatial dimension in mm.
-        If grid_spacing is None, MRIs won't be resliced. Note that in this case
-        both volumes must have the same number of slices in every
-        spatial dimension.
+    spacing : tuple | int | float | list | None
+        If morphing VolSourceEstimate, spacing is a tuple carrying the voxel
+        size of volume for each spatial dimension in mm.
+        If spacing is None, MRIs won't be resliced. Note that in this case
+        both volumes (used to compute the morph) must have the same number of
+        spatial dimensions.
+        If morphing SourceEstimate or Vector SourceEstimate, spacing can be
+        int, list (of two arrays), or None, defining the resolution of the
+        icosahedral mesh (typically 5). If None, all vertices will be used
+        (potentially filling the surface). If a list, then values will be
+        morphed to the set of vertices specified in in spacing[0] and
+        spacing[1].
+        Note that specifying the vertices (e.g., spacing=[np.arange(10242),
+        np.arange(10242)] for fsaverage on a standard spacing 5 source space)
+        can be substantially faster than computing vertex locations.
     smooth : int | None
         Number of iterations for the smoothing of the surface data.
         If None, smooth is automatically defined to fill the surface
@@ -123,7 +148,8 @@ class SourceMorph(object):
 
     def __init__(self, src, subject_from=None, subject_to='fsaverage',
                  subjects_dir=None, niter_affine=(100, 100, 10),
-                 niter_sdr=(5, 5, 3), grid_spacing=(5., 5., 5.), smooth=None,
+                 niter_sdr=(5, 5, 3), spacing=5,
+                 smooth=None,
                  warn=True, xhemi=False, verbose=False):
 
         # it's impossible to use the class without passing this check, so it
@@ -140,11 +166,11 @@ class SourceMorph(object):
         self.subjects_dir = subjects_dir
         self.niter_affine = niter_affine
         self.niter_sdr = niter_sdr
-        self.grid_spacing = grid_spacing
+        self.spacing = spacing
         self.smooth = smooth
         self.warn = warn
         self.xhemi = xhemi
-        self.morph_data = None
+        self.morph_data = dict()
 
         if src is None:
             return
@@ -162,10 +188,10 @@ class SourceMorph(object):
             self.subject_from = src[0]['subject_his_id']
 
         if self.subject_from is None:
-            raise KeyError('subject id in src is None. Please specify'
+            raise KeyError('subject id in src is None. Please specify '
                            'subject_from')
-
-        self.morph_data = _compute_morph_data(self, src, verbose=verbose)
+        self.morph_data.update(_get_src_data(src))
+        self.morph_data.update(_compute_morph_data(self, verbose=verbose))
 
     # Forward verbose decorator to _apply_morph_data
     def __call__(self, stc_from, verbose=None):
@@ -185,15 +211,12 @@ class SourceMorph(object):
         stc_to : VolSourceEstimate | SourceEstimate | VectorSourceEstimate
             The morphed source estimate.
         """
+        if stc_from.subject is None:
+            stc_from.subject = self.subject_from
+
         if stc_from.subject != self.subject_from:
             raise ValueError('stc_from.subject and '
                              'morph.subject_from must match')
-
-        # if VolSourceEstimate update mri properties for '.as_volume'
-        if self.kind == 'volume':
-            self.morph_data.update(
-                {'mri_zooms': self.morph_data['mri_zooms_to'],
-                 'mri_shape': self.morph_data['mri_shape_to']})
 
         return _apply_morph_data(self, stc_from, verbose=verbose)
 
@@ -207,7 +230,7 @@ class SourceMorph(object):
         if self.kind == 'volume':
             s += ", niter_affine : {}".format(self.niter_affine)
             s += ", niter_sdr : {}".format(self.niter_sdr)
-            s += ", grid_spacing : {}".format(self.grid_spacing)
+            s += ", spacing : {}".format(self.spacing)
 
         elif self.kind == 'surface' or self.kind == 'vector':
             s += ", smooth : %s" % self.smooth
@@ -238,66 +261,37 @@ class SourceMorph(object):
                    overwrite=True)
         logger.info('[done]')
 
-    def as_volume(self, stc, fname=None, mri_resolution=False):
+    def as_volume(self, stc, fname=None, mri_resolution=False, mri_space=True):
         """Return volume source space as Nifti1Image and/or save to disk.
 
         Parameters
         ----------
-        stc : VolSourceEstimate | SourceEstimate | VectorSourceEstimate
+        stc : VolSourceEstimate
             Data to be transformed
         fname : str | None
             String to where to save the volume. If not None that volume will
             be saved at fname.
-        mri_resolution : bool | tuple
+        mri_resolution: bool | tuple | int | float
             Whether to use MRI resolution. If False the morph's resolution
             will be used. If tuple the voxel size must be given in float values
             in mm. E.g. mri_resolution=(3., 3., 3.)
+            WARNING: if you have many time points the file produced can be
+            huge.
+        mri_space : bool
+            Whether the image to world registration should be in mri space.
 
         Returns
         -------
         img : instance of Nifti1Image
             The image object.
         """
-        import nibabel as nib
-        from dipy.align.reslice import reslice
-
-        shape = tuple([int(i) for i in self.morph_data['morph_shape']])
-        affine = self.morph_data['morph_affine']
-        zooms = self.morph_data['morph_zooms'][:3]
-
-        hdr = nib.nifti1.Nifti1Header()
-        hdr.set_xyzt_units('mm', 'msec')
-        hdr['pixdim'][4] = 1e3 * stc.tstep
-
-        new_zooms = None
-
-        if isinstance(mri_resolution, bool) and mri_resolution:
-            new_zooms = self.morph_data['mri_zooms']
-
-        elif isinstance(mri_resolution, tuple):
-            new_zooms = mri_resolution
-
-        img = np.zeros(shape).reshape(-1, stc.shape[1])
-        img[stc.vertices, :] = stc.data
-
-        img = img.reshape(shape + (-1,))
-
-        img = nib.Nifti1Image(img, affine, header=hdr)
-
-        if new_zooms is not None:
-            new_zooms = new_zooms[:3]
-            img, affine = reslice(img.get_data(),
-                                  img.affine,
-                                  zooms,
-                                  new_zooms)
-            img = nib.Nifti1Image(img, affine)
-            zooms = new_zooms
-        img.header.set_zooms(zooms + (1,))
-        if fname is not None:
-            nib.save(img, fname)
-        return img
+        return _stc_as_volume(self, stc, fname=fname,
+                              mri_resolution=mri_resolution,
+                              mri_space=mri_space)
 
 
+###############################################################################
+# I/O
 @verbose
 def read_source_morph(fname, verbose=None):
     """Load the morph for source estimates from a file.
@@ -323,33 +317,155 @@ def read_source_morph(fname, verbose=None):
         raise IOError('cannot read file: %s' % fname)
 
     source_morph = SourceMorph(None)
-    source_morph.__dict__ = morph_data
+    source_morph.__dict__.update(morph_data)
     logger.info('[done]')
     return source_morph
 
 
-def _compute_morph_data(morph, src, verbose=None):
-    """Compute source estimate specific morph."""
+###############################################################################
+# Helper functions for SourceMorph methods
+def _stc_as_volume(morph, stc, fname=None, mri_resolution=False,
+                   mri_space=True):
+    """Return volume source space as Nifti1Image and/or save to disk.
 
-    morph_data = []
+    Parameters
+    ----------
+    morph : instance of SourceMorph
+        Instance of SourceMorph carrying the relevant volume transform
+        information. Typically computed using SourceMorph()
+    stc : VolSourceEstimate
+        Data to be transformed
+    fname : str | None
+        String to where to save the volume. If not None that volume will
+        be saved at fname.
+    mri_resolution: bool | tuple | int | float
+        Whether to use MRI resolution. If False the morph's resolution
+        will be used. If tuple the voxel size must be given in float values
+        in mm. E.g. mri_resolution=(3., 3., 3.)
+        WARNING: if you have many time points the file produced can be
+        huge.
+    mri_space : bool
+        Whether the image to world registration should be in mri space.
+
+    Returns
+    -------
+    img : instance of Nifti1Image
+        The image object.
+    """
+    import nibabel as nib
+    from dipy.align.reslice import reslice
+
+    if not isinstance(stc, VolSourceEstimate):
+        raise ValueError('Only volume source estimates can be converted to '
+                         'volumes')
+
+    # this is a special case when as_volume is called without having done a
+    # morph beforehand (to assure compatibility to previous versions)
+    if 'morph_shape' not in morph.morph_data:
+        img = _interpolate_data(stc, morph, mri_resolution=mri_resolution,
+                                mri_space=mri_space)
+        if fname is not None:
+            nib.save(img, fname)
+        return img
+
+    new_zooms = None
+
+    if isinstance(mri_resolution, bool) and mri_resolution:
+        new_zooms = _get_zooms_orig(morph)
+
+    if isinstance(mri_resolution, (int, float)) and not isinstance(
+            mri_resolution, bool):
+        # use iso voxel size
+        new_zooms = (
+            float(mri_resolution), float(mri_resolution),
+            float(mri_resolution))
+
+    if isinstance(mri_resolution, tuple):
+        new_zooms = mri_resolution
+
+    shape = tuple([int(i) for i in morph.morph_data['morph_shape']])
+    affine = morph.morph_data['morph_affine']
+    zooms = morph.morph_data['morph_zooms'][:3]
+
+    hdr = nib.nifti1.Nifti1Header()
+    hdr.set_xyzt_units('mm', 'msec')
+    hdr['pixdim'][4] = 1e3 * stc.tstep
+
+    img = np.zeros(shape).reshape(-1, stc.shape[1])
+    img[stc.vertices, :] = stc.data
+
+    img = img.reshape(shape + (-1,))
+
+    img = nib.Nifti1Image(img, affine, header=hdr)
+
+    if new_zooms is not None:
+        new_zooms = new_zooms[:3]
+        img, affine = reslice(img.get_data(),
+                              img.affine,
+                              zooms,
+                              new_zooms)
+        img = nib.Nifti1Image(img, affine)
+        zooms = new_zooms
+    img.header.set_zooms(tuple(zooms) + (1,))
+    if fname is not None:
+        nib.save(img, fname)
+    return img
+
+
+def _get_src_data(src):
+    """Obtain src data relevant for as_volume() and __call__()."""
+    src_data = dict()
+    src_t = copy.deepcopy(src)
+    if src.kind == 'volume':
+        shape = src_t[0]['shape']
+        src_data.update({'src_shape': (shape[2], shape[1], shape[0]),
+                         'src_affine_vox': src_t[0]['vox_mri_t']['trans'],
+                         'src_affine_src': src_t[0]['src_mri_t']['trans'],
+                         'src_affine_ras': src_t[0]['mri_ras_t']['trans'],
+                         'src_shape_full': (
+                             src_t[0]['mri_height'], src_t[0]['mri_depth'],
+                             src_t[0]['mri_width']),
+                         'interpolator': src_t[0]['interpolator'],
+                         'inuse': src_t[0]['inuse']})
+    elif src_t.kind == 'surface':
+        hemis = []
+        for n, s in enumerate(src_t):
+            src_data.update({str(n): s['vertno']})
+            hemis.append(n)
+        src_data.update({'hemis': hemis})
+    del src_t
+    return src_data
+
+
+def _compute_morph_data(morph, verbose=None):
+    """Compute source estimate specific morph."""
+    morph_data = dict()
+    morph_data.update(morph.morph_data)
+    subjects_dir = get_subjects_dir(morph.subjects_dir,
+                                    raise_error=True)
+
+    # get data from src - it is necessary to split this up (src info and
+    # morphing), because in order to provide backwards compatibility it should
+    # be possible to make a volume from an un-morphed stc. Hence in that case a
+    # morpher would be created that only holds src data needed for as_volume.
+    # Note further that this is only internal API.
 
     # VolSourceEstimate
-    if morph.kind == 'volume':
+    if morph.kind == 'volume' and morph.subject_to is not None:
 
         logger.info('volume source space inferred...')
-
         import nibabel as nib
 
         # load moving mri
         mri_subpath = op.join('mri', 'brain.mgz')
-        mri_path_from = op.join(morph.subjects_dir, morph.subject_from,
+        mri_path_from = op.join(subjects_dir, morph.subject_from,
                                 mri_subpath)
 
         logger.info('loading %s as moving volume' % mri_path_from)
         mri_from = nib.load(mri_path_from)
 
         # load static mri
-        static_path = op.join(morph.subjects_dir, morph.subject_to)
+        static_path = op.join(subjects_dir, morph.subject_to)
 
         if not op.isdir(static_path):
             mri_path_to = static_path
@@ -362,29 +478,14 @@ def _compute_morph_data(morph, src, verbose=None):
             raise IOError('cannot read file: %s' % mri_path_to)
 
         # pre-compute non-linear morph
-        morph_data = _compute_morph_sdr(
+        morph_data.update(_compute_morph_sdr(
             mri_from,
             mri_to,
             niter_affine=morph.niter_affine,
             niter_sdr=morph.niter_sdr,
-            grid_spacing=morph.grid_spacing,
-            verbose=verbose)
+            spacing=morph.spacing,
+            verbose=verbose))
 
-        # get MRI and source space meta data
-        morph_data['mri_shape'] = mri_from.shape
-        morph_data['mri_zooms'] = mri_from.header.get_zooms()
-
-        morph_data['mri_shape_to'] = mri_to.shape
-        morph_data['mri_zooms_to'] = mri_to.header.get_zooms()
-
-        morph_data['src_shape_full'] = mri_from.shape
-        shape = src[0]['shape']
-        morph_data['src_shape'] = (shape[2], shape[1], shape[0])
-        morph_data['src_affine'] = np.dot(src[0]['mri_ras_t']['trans'],
-                                          src[0]['vox_mri_t']['trans'])
-
-        morph_data['interpolator'] = src[0]['interpolator']
-        morph_data['inuse'] = src[0]['inuse']
 
     # SourceEstimate | VectorSourceEstimate
     elif morph.kind == 'surface':
@@ -392,31 +493,41 @@ def _compute_morph_data(morph, src, verbose=None):
 
         # get surface data
         data_from = []
-        for s in src:
-            data_from.append(s['vertno'])
+        hemis = morph.morph_data['hemis']
+        for h in hemis:
+            data_from.append(morph.morph_data[str(h)])
 
         data_to = None
         if morph.subject_to == 'fsaverage':
             data_to = [np.arange(10242)] * len(data_from)
 
+        if isinstance(morph.spacing, int) or isinstance(morph.spacing, list):
+            data_to = grade_to_vertices(morph.subject_to, morph.spacing,
+                                        subjects_dir, 1)
+
+        if data_to is None:
+            raise ValueError('Please specify target to morph to.')
+
         # pre-compute morph matrix
-        morph_data = compute_morph_matrix(
+        morph_mat = _compute_morph_matrix(
             subject_from=morph.subject_from,
             subject_to=morph.subject_to,
             vertices_from=data_from,
             vertices_to=data_to,
-            subjects_dir=morph.subjects_dir,
+            subjects_dir=subjects_dir,
             smooth=morph.smooth,
             warn=morph.warn,
             xhemi=morph.xhemi,
             verbose=verbose)
-        morph_data = {'morph_mat': morph_data, 'vertno': data_to}
+        morph_data.update({'morph_mat': morph_mat, 'vertno': data_to})
     return morph_data
 
 
-def _interpolate_data(stc, morph):
+def _interpolate_data(stc, morph, mri_resolution=True,
+                      mri_space=True):
     """Interpolate source estimate data to MRI."""
     import nibabel as nib
+    from dipy.align.reslice import reslice
 
     n_times = stc.data.shape[1]
     shape = morph.morph_data['src_shape']
@@ -424,11 +535,31 @@ def _interpolate_data(stc, morph):
     shape = (n_times, shape[0], shape[1], shape[2])
     vol = np.zeros(shape)
 
-    mri_shape3d = morph.morph_data['src_shape_full']
-    mri_shape = (n_times, mri_shape3d[0], mri_shape3d[1],
-                 mri_shape3d[2])
-    mri_vol = np.zeros(mri_shape)
-    interpolator = morph.morph_data['interpolator']
+    voxel_size_defined = False
+
+    if isinstance(mri_resolution, (int, float)) and not isinstance(
+            mri_resolution, bool):
+        # use iso voxel size
+        mri_resolution = (
+            float(mri_resolution), float(mri_resolution),
+            float(mri_resolution))
+
+    if isinstance(mri_resolution, tuple):
+        voxel_size = mri_resolution
+        voxel_size_defined = True
+        mri_resolution = True
+
+    if 'morph_shape' not in morph.morph_data and voxel_size_defined:
+        raise ValueError(
+            "Cannot infer original voxel size for reslicing... "
+            "set mri_resolution to boolean value or apply morph first.")
+
+    if mri_resolution:
+        mri_shape3d = morph.morph_data['src_shape_full']
+        mri_shape = (n_times, mri_shape3d[0], mri_shape3d[1],
+                     mri_shape3d[2])
+        mri_vol = np.zeros(mri_shape)
+        interpolator = morph.morph_data['interpolator']
 
     mask3d = morph.morph_data['inuse'].reshape(shape3d).astype(np.bool)
     n_vertices = np.sum(mask3d)
@@ -440,13 +571,21 @@ def _interpolate_data(stc, morph):
 
     n_vertices_seen += n_vertices
 
-    for k, v in enumerate(vol):
-        mri_vol[k] = (interpolator * v.ravel()).reshape(mri_shape3d)
-    vol = mri_vol
+    if mri_resolution:
+        for k, v in enumerate(vol):
+            mri_vol[k] = (interpolator * v.ravel()).reshape(mri_shape3d)
+        vol = mri_vol
 
     vol = vol.T
 
-    affine = morph.morph_data['src_affine']
+    affine = morph.morph_data['src_affine_vox']
+
+    if not mri_resolution:
+        affine = morph.morph_data['src_affine_src']
+
+    if mri_space:
+        affine = np.dot(morph.morph_data['src_affine_ras'], affine)
+
     affine[:3] *= 1e3
 
     header = nib.nifti1.Nifti1Header()
@@ -456,17 +595,26 @@ def _interpolate_data(stc, morph):
     with warnings.catch_warnings(record=True):  # nibabel<->numpy warning
         img = nib.Nifti1Image(vol, affine, header=header)
 
+    if voxel_size_defined:
+        # reslice mri
+        img, img_affine = reslice(
+            img.get_data(),
+            img.affine,
+            _get_zooms_orig(morph),
+            voxel_size)
+
+        img = nib.Nifti1Image(img, img_affine, header=header)
+
     return img
 
 
 ###############################################################################
-# Morph for VolSourceEstimate -
-# compute AffineMap and DiffeomorphicMap from MRIs
+# Morph for VolSourceEstimate
 @verbose
 def _compute_morph_sdr(mri_from, mri_to,
                        niter_affine=(100, 100, 10),
                        niter_sdr=(5, 5, 3),
-                       grid_spacing=(5., 5., 5.),
+                       spacing=(5., 5., 5.),
                        verbose=None):
     """Get a matrix that morphs data from one subject to another.
 
@@ -486,9 +634,10 @@ def _compute_morph_sdr(mri_from, mri_to,
         iterations to perform the sdr transform. Increasing index values
         for the tuple mean later levels and each int represents the number
         of iterations in that level.
-    grid_spacing : tuple
-        Voxel size of volume for each spatial dimension.
-        If grid_spacing is None, MRIs won't be resliced. Note that in this case
+    spacing : tuple | int | float | None
+        Voxel size of volume for each spatial dimension separately (tuple) or
+        isometric (int).
+        If spacing is None, MRIs won't be resliced. Note that in this case
         both volumes must have the same number of slices in every
         spatial dimension.
     verbose : bool | str | int | None
@@ -536,15 +685,19 @@ def _compute_morph_sdr(mri_from, mri_to,
     morph = dict()
 
     # use voxel size of mri_from
-    if grid_spacing is None:
-        grid_spacing = mri_from.header.get_zooms()[:3]
+    if spacing is None:
+        spacing = mri_from.header.get_zooms()[:3]
+
+    # use iso voxel size
+    if isinstance(spacing, (int, float)):
+        spacing = (float(spacing), float(spacing), float(spacing))
 
     # reslice mri_from
     mri_from_res, mri_from_res_affine = reslice(
         mri_from.get_data(),
         mri_from.affine,
         mri_from.header.get_zooms()[:3],
-        grid_spacing)
+        spacing)
 
     mri_from = nib.Nifti1Image(mri_from_res, mri_from_res_affine)
 
@@ -553,7 +706,7 @@ def _compute_morph_sdr(mri_from, mri_to,
         mri_to.get_data(),
         mri_to.affine,
         mri_to.header.get_zooms()[:3],
-        grid_spacing)
+        spacing)
 
     mri_to = nib.Nifti1Image(mri_to_res, mri_to_res_affine)
 
@@ -616,17 +769,413 @@ def _compute_morph_sdr(mri_from, mri_to,
 
     # compute mapping
     mapping = sdr.optimize(mri_to, mri_from_affine)
-
-    morph['morph_shape'] = tuple(mapping.domain_shape.astype('float'))
-    morph['morph_zooms'] = grid_spacing
-    morph['morph_affine'] = mri_to_grid2world
-
-    morph['AffineMap'] = affine.__dict__
-    morph['DiffeomorphicMap'] = mapping.__dict__
+    morph.update({'morph_shape': tuple(mapping.domain_shape.astype('float')),
+                  'morph_zooms': spacing,
+                  'morph_affine': mri_to_grid2world,
+                  'AffineMap': affine.__dict__,
+                  'DiffeomorphicMap': mapping.__dict__})
 
     logger.info('done.')
 
     return morph
+
+
+###############################################################################
+# Morph for SourceEstimate |  VectorSourceEstimate
+@deprecated("This function is deprecated and might be removed in a future "
+            "release. Use morph = SourceMorph(src_from) and "
+            "stc_to = morph(stc_from)")
+def compute_morph_matrix(subject_from, subject_to, vertices_from, vertices_to,
+                         smooth=None, subjects_dir=None, warn=True,
+                         xhemi=False, verbose=None):
+    """Wrapper for _compute_morph_matrix to assure backwards compatibility."""
+    return _compute_morph_matrix(subject_from, subject_to, vertices_from,
+                                 vertices_to, smooth, subjects_dir, warn,
+                                 xhemi, verbose)
+
+
+@verbose
+def _compute_morph_matrix(subject_from, subject_to, vertices_from, vertices_to,
+                          smooth=None, subjects_dir=None, warn=True,
+                          xhemi=False, verbose=None):
+    """Get a matrix that morphs data from one subject to another.
+
+    Parameters
+    ----------
+    subject_from : str
+        Name of the original subject as named in the SUBJECTS_DIR.
+    subject_to : str
+        Name of the subject on which to morph as named in the SUBJECTS_DIR.
+    vertices_from : list of arrays of int
+        Vertices for each hemisphere (LH, RH) for subject_from.
+    vertices_to : list of arrays of int
+        Vertices for each hemisphere (LH, RH) for subject_to.
+    smooth : int or None
+        Number of iterations for the smoothing of the surface data.
+        If None, smooth is automatically defined to fill the surface
+        with non-zero values.
+    subjects_dir : str
+        Path to SUBJECTS_DIR is not set in the environment.
+    warn : bool
+        If True, warn if not all vertices were used.
+    xhemi : bool
+        Morph across hemisphere. Currently only implemented for
+        ``subject_to == subject_from``. See notes below.
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see :func:`mne.verbose`
+        and :ref:`Logging documentation <tut_logging>` for more).
+
+    Returns
+    -------
+    morph_matrix : sparse matrix
+        matrix that morphs data from ``subject_from`` to ``subject_to``.
+
+    Notes
+    -----
+    This function can be used to morph data between hemispheres by setting
+    ``xhemi=True``. The full cross-hemisphere morph matrix maps left to right
+    and right to left. A matrix for cross-mapping only one hemisphere can be
+    constructed by specifying the appropriate vertices, for example, to map the
+    right hemisphere to the left:
+    ``vertices_from=[[], vert_rh], vertices_to=[vert_lh, []]``.
+
+    Cross-hemisphere mapping requires appropriate ``sphere.left_right``
+    morph-maps in the subject's directory. These morph maps are included
+    with the ``fsaverage_sym`` FreeSurfer subject, and can be created for other
+    subjects with the ``mris_left_right_register`` FreeSurfer command. The
+    ``fsaverage_sym`` subject is included with FreeSurfer > 5.1 and can be
+    obtained as described `here
+    <http://surfer.nmr.mgh.harvard.edu/fswiki/Xhemi>`_. For statistical
+    comparisons between hemispheres, use of the symmetric ``fsaverage_sym``
+    model is recommended to minimize bias [1]_.
+
+    References
+    ----------
+    .. [1] Greve D. N., Van der Haegen L., Cai Q., Stufflebeam S., Sabuncu M.
+           R., Fischl B., Brysbaert M.
+           A Surface-based Analysis of Language Lateralization and Cortical
+           Asymmetry. Journal of Cognitive Neuroscience 25(9), 1477-1492, 2013.
+    """
+    logger.info('Computing morph matrix...')
+    subjects_dir = get_subjects_dir(subjects_dir, raise_error=True)
+
+    tris = _get_subject_sphere_tris(subject_from, subjects_dir)
+    maps = read_morph_map(subject_from, subject_to, subjects_dir, xhemi)
+
+    # morph the data
+
+    if xhemi:
+        hemi_indexes = [(0, 1), (1, 0)]
+    else:
+        hemi_indexes = [(0, 0), (1, 1)]
+    morpher = []
+    for hemi_from, hemi_to in hemi_indexes:
+        idx_use = vertices_from[hemi_from]
+        if len(idx_use) == 0:
+            continue
+        e = mesh_edges(tris[hemi_from])
+        e.data[e.data == 2] = 1
+        n_vertices = e.shape[0]
+        e = e + sparse.eye(n_vertices, n_vertices)
+        m = sparse.eye(len(idx_use), len(idx_use), format='csr')
+        mm = _morph_buffer(m, idx_use, e, smooth, n_vertices,
+                           vertices_to[hemi_to], maps[hemi_from], warn=warn)
+        morpher.append(mm)
+
+    if len(morpher) == 0:
+        raise ValueError("Empty morph-matrix")
+    elif len(morpher) == 1:
+        morpher = morpher[0]
+    else:
+        morpher = sparse_block_diag(morpher, format='csr')
+    logger.info('[done]')
+    return morpher
+
+
+@verbose
+def grade_to_vertices(subject, grade, subjects_dir=None, n_jobs=1,
+                      verbose=None):
+    """Convert a grade to source space vertices for a given subject.
+
+    Parameters
+    ----------
+    subject : str
+        Name of the subject
+    grade : int | list
+        Resolution of the icosahedral mesh (typically 5). If None, all
+        vertices will be used (potentially filling the surface). If a list,
+        then values will be morphed to the set of vertices specified in
+        in grade[0] and grade[1]. Note that specifying the vertices (e.g.,
+        grade=[np.arange(10242), np.arange(10242)] for fsaverage on a
+        standard grade 5 source space) can be substantially faster than
+        computing vertex locations. Note that if subject='fsaverage'
+        and 'grade=5', this set of vertices will automatically be used
+        (instead of computed) for speed, since this is a common morph.
+    subjects_dir : str | None
+        Path to SUBJECTS_DIR if it is not set in the environment
+    n_jobs : int
+        Number of jobs to run in parallel
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see :func:`mne.verbose`
+        and :ref:`Logging documentation <tut_logging>` for more).
+
+    Returns
+    -------
+    vertices : list of arrays of int
+        Vertex numbers for LH and RH
+    """
+    # add special case for fsaverage for speed
+    if subject == 'fsaverage' and grade == 5:
+        return [np.arange(10242), np.arange(10242)]
+    subjects_dir = get_subjects_dir(subjects_dir, raise_error=True)
+
+    spheres_to = [op.join(subjects_dir, subject, 'surf',
+                          xh + '.sphere.reg') for xh in ['lh', 'rh']]
+    lhs, rhs = [read_surface(s)[0] for s in spheres_to]
+
+    if grade is not None:  # fill a subset of vertices
+        if isinstance(grade, list):
+            if not len(grade) == 2:
+                raise ValueError('grade as a list must have two elements '
+                                 '(arrays of output vertices)')
+            vertices = grade
+        else:
+            # find which vertices to use in "to mesh"
+            ico = _get_ico_tris(grade, return_surf=True)
+            lhs /= np.sqrt(np.sum(lhs ** 2, axis=1))[:, None]
+            rhs /= np.sqrt(np.sum(rhs ** 2, axis=1))[:, None]
+
+            # Compute nearest vertices in high dim mesh
+            parallel, my_compute_nearest, _ = \
+                parallel_func(_compute_nearest, n_jobs)
+            lhs, rhs, rr = [a.astype(np.float32)
+                            for a in [lhs, rhs, ico['rr']]]
+            vertices = parallel(my_compute_nearest(xhs, rr)
+                                for xhs in [lhs, rhs])
+            # Make sure the vertices are ordered
+            vertices = [np.sort(verts) for verts in vertices]
+            for verts in vertices:
+                if (np.diff(verts) == 0).any():
+                    raise ValueError(
+                        'Cannot use icosahedral grade %s with subject %s, '
+                        'mapping %s vertices onto the high-resolution mesh '
+                        'yields repeated vertices, use a lower grade or a '
+                        'list of vertices from an existing source space'
+                        % (grade, subject, len(verts)))
+    else:  # potentially fill the surface
+        vertices = [np.arange(lhs.shape[0]), np.arange(rhs.shape[0])]
+
+    return vertices
+
+
+@verbose
+def _morph_buffer(data, idx_use, e, smooth, n_vertices, nearest, maps,
+                  warn=True, verbose=None):
+    """Morph data from one subject's source space to another.
+
+    Parameters
+    ----------
+    data : array, or csr sparse matrix
+        A n_vertices [x 3] x n_times (or other dimension) dataset to morph.
+    idx_use : array of int
+        Vertices from the original subject's data.
+    e : sparse matrix
+        The mesh edges of the "from" subject.
+    smooth : int
+        Number of smoothing iterations to perform. A hard limit of 100 is
+        also imposed.
+    n_vertices : int
+        Number of vertices.
+    nearest : array of int
+        Vertices on the destination surface to use.
+    maps : sparse matrix
+        Morph map from one subject to the other.
+    warn : bool
+        If True, warn if not all vertices were used.
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see :func:`mne.verbose`
+        and :ref:`Logging documentation <tut_logging>` for more).
+
+    Returns
+    -------
+    data_morphed : array, or csr sparse matrix
+        The morphed data (same type as input).
+    """
+    # When operating on vector data, morph each dimension separately
+    if data.ndim == 3:
+        data_morphed = np.zeros((len(nearest), 3, data.shape[2]),
+                                dtype=data.dtype)
+        for dim in range(3):
+            data_morphed[:, dim, :] = _morph_buffer(
+                data=data[:, dim, :], idx_use=idx_use, e=e, smooth=smooth,
+                n_vertices=n_vertices, nearest=nearest, maps=maps, warn=warn,
+                verbose=verbose
+            )
+        return data_morphed
+
+    n_iter = 99  # max nb of smoothing iterations (minus one)
+    if smooth is not None:
+        if smooth <= 0:
+            raise ValueError('The number of smoothing operations ("smooth") '
+                             'has to be at least 1.')
+        smooth -= 1
+    # make sure we're in CSR format
+    e = e.tocsr()
+    if sparse.issparse(data):
+        use_sparse = True
+        if not isinstance(data, sparse.csr_matrix):
+            data = data.tocsr()
+    else:
+        use_sparse = False
+
+    done = False
+    # do the smoothing
+    for k in range(n_iter + 1):
+        # get the row sum
+        mult = np.zeros(e.shape[1])
+        mult[idx_use] = 1
+        idx_use_data = idx_use
+        data_sum = e * mult
+
+        # new indices are non-zero sums
+        idx_use = np.where(data_sum)[0]
+
+        # typically want to make the next iteration have these indices
+        idx_out = idx_use
+
+        # figure out if this is the last iteration
+        if smooth is None:
+            if k == n_iter or len(idx_use) >= n_vertices:
+                # stop when vertices filled
+                idx_out = None
+                done = True
+        elif k == smooth:
+            idx_out = None
+            done = True
+
+        # do standard smoothing multiplication
+        data = _morph_mult(data, e, use_sparse, idx_use_data, idx_out)
+
+        if done is True:
+            break
+
+        # do standard normalization
+        if use_sparse:
+            data.data /= data_sum[idx_use].repeat(np.diff(data.indptr))
+        else:
+            data /= data_sum[idx_use][:, None]
+
+    # do special normalization for last iteration
+    if use_sparse:
+        data_sum[data_sum == 0] = 1
+        data.data /= data_sum.repeat(np.diff(data.indptr))
+    else:
+        data[idx_use, :] /= data_sum[idx_use][:, None]
+    if len(idx_use) != len(data_sum) and warn:
+        warn_('%s/%s vertices not included in smoothing, consider increasing '
+              'the number of steps'
+              % (len(data_sum) - len(idx_use), len(data_sum)))
+
+    logger.info('    %d smooth iterations done.' % (k + 1))
+
+    data_morphed = maps[nearest, :] * data
+    return data_morphed
+
+
+def _morph_mult(data, e, use_sparse, idx_use_data, idx_use_out=None):
+    """Help morphing.
+
+    Equivalent to "data = (e[:, idx_use_data] * data)[idx_use_out]"
+    but faster.
+    """
+    if len(idx_use_data) < e.shape[1]:
+        if use_sparse:
+            data = e[:, idx_use_data] * data
+        else:
+            # constructing a new sparse matrix is faster than sub-indexing
+            # e[:, idx_use_data]!
+            col, row = np.meshgrid(np.arange(data.shape[1]), idx_use_data)
+            d_sparse = sparse.csr_matrix((data.ravel(),
+                                          (row.ravel(), col.ravel())),
+                                         shape=(e.shape[1], data.shape[1]))
+            data = e * d_sparse
+            data = np.asarray(data.todense())
+    else:
+        data = e * data
+
+    # trim data
+    if idx_use_out is not None:
+        data = data[idx_use_out]
+    return data
+
+
+def _sparse_argmax_nnz_row(csr_mat):
+    """Return index of the maximum non-zero index in each row."""
+    n_rows = csr_mat.shape[0]
+    idx = np.empty(n_rows, dtype=np.int)
+    for k in range(n_rows):
+        row = csr_mat[k].tocoo()
+        idx[k] = row.col[np.argmax(row.data)]
+    return idx
+
+
+def _morph_sparse(stc, subject_from, subject_to, subjects_dir=None):
+    """Morph sparse source estimates to an other subject.
+
+    Parameters
+    ----------
+    stc : SourceEstimate | VectorSourceEstimate
+        The sparse STC.
+    subject_from : str
+        The subject on which stc is defined.
+    subject_to : str
+        The target subject.
+    subjects_dir : str
+        Path to SUBJECTS_DIR if it is not set in the environment.
+
+    Returns
+    -------
+    stc_morph : SourceEstimate | VectorSourceEstimate
+        The morphed source estimates.
+    """
+    maps = read_morph_map(subject_to, subject_from, subjects_dir)
+    stc_morph = stc.copy()
+    stc_morph.subject = subject_to
+
+    cnt = 0
+    for k, hemi in enumerate(['lh', 'rh']):
+        if stc.vertices[k].size > 0:
+            map_hemi = maps[k]
+            vertno_k = _sparse_argmax_nnz_row(map_hemi[stc.vertices[k]])
+            order = np.argsort(vertno_k)
+            n_active_hemi = len(vertno_k)
+            data_hemi = stc_morph.data[cnt:cnt + n_active_hemi]
+            stc_morph.data[cnt:cnt + n_active_hemi] = data_hemi[order]
+            stc_morph.vertices[k] = vertno_k[order]
+            cnt += n_active_hemi
+        else:
+            stc_morph.vertices[k] = np.array([], int)
+
+    return stc_morph
+
+
+def _get_subject_sphere_tris(subject, subjects_dir):
+    spheres = [op.join(subjects_dir, subject, 'surf',
+                       xh + '.sphere.reg') for xh in ['lh', 'rh']]
+    tris = [read_surface(s)[1] for s in spheres]
+    return tris
+
+
+###############################################################################
+# Apply morph to source estimate
+def _get_zooms_orig(morph):
+    """Compute src zooms from morph zooms, morph shape and src shape."""
+    morph_zooms = morph.morph_data['morph_zooms']
+    morph_shape = morph.morph_data['morph_shape']
+    src_shape = morph.morph_data['src_shape_full']
+
+    return [mz / ss * ms for mz, ms, ss in
+            zip(morph_zooms, morph_shape, src_shape)]
 
 
 @verbose
@@ -654,7 +1203,8 @@ def _apply_morph_data(morph, stc_from, verbose=None):
         from dipy.align.reslice import reslice
 
         # prepare data to be morphed
-        img_to = _interpolate_data(stc_from, morph)
+        img_to = _interpolate_data(stc_from, morph, mri_resolution=True,
+                                   mri_space=True)
 
         # setup morphs
         affine_morph = AffineMap(None)
@@ -667,7 +1217,7 @@ def _apply_morph_data(morph, stc_from, verbose=None):
         img_to, img_to_affine = reslice(
             img_to.get_data(),
             morph.morph_data['morph_affine'],
-            morph.morph_data['mri_zooms'],
+            _get_zooms_orig(morph),
             morph.morph_data['morph_zooms'])
 
         # morph data
